fix: 오늘인지 판단하는 로직 변경
