+++ frontend/src/routes/_bombom/today.tsx
@@ -16,7 +16,7 @@ import {
 } from '@/pages/guide-detail/constants/guideMail';
 import ArticleCardListSkeleton from '@/pages/today/components/ArticleCardList/ArticleCardListSkeleton';
 import { theme } from '@/styles/theme';
-import { formatDate, isToday } from '@/utils/date';
+import { formatDate } from '@/utils/date';
 import { createStorage } from '@/utils/localStorage';
 import type { Device } from '@/hooks/useDevice';
 import type { LocalGuideMail } from '@/types/guide';
@@ -60,7 +60,7 @@ function Index() {
   const guideMailReadMailIds = guideMails?.readMailIds ?? [];
 
   const guideArticles =
-    guideMails?.createdAt && isToday(new Date(guideMails?.createdAt))
+    guideMails?.createdAt && guideMails.createdAt === formatDate(new Date())
       ? GUIDE_MAILS
       : [];
